Agrego Lombok

+++ src/main/java/com/accenture/granja/services/GranjaService.java
@@ -1,12 +1,10 @@
 package com.accenture.granja.services;
 
 import java.util.List;
-
 import javax.transaction.Transactional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
+import com.accenture.granja.exceptions.NoContentException;
 import com.accenture.granja.model.Granja;
 import com.accenture.granja.repository.GranjaRepository;
 @Service
@@ -15,27 +13,41 @@ public class GranjaService {
 	@Autowired
 	private GranjaRepository granjaRepo;
 
-	public Granja buscarGranja(Long id) {
-		Granja granja= granjaRepo.findById(id).orElse(null);
-		return granja;
+	public List<Granja> buscarGranjas() {
+		List<Granja> granjas = granjaRepo.findAll();
+        if (granjas.isEmpty()) {
+            throw new NoContentException("No hay granjas por el momento");
+        }
+        return granjas;
 	}
 	
-	public List<Granja> buscarGranjas() {
-		List<Granja> granjas= granjaRepo.findAll();
-		return granjas;
-	}
-
-	public Granja editarGranja(Granja granja) {
-		return granjaRepo.save(granja);
-	}
-
-
+	public Granja buscarGranja(Long id) {
+		 return granjaRepo.findById(id)
+	                .orElseThrow(() -> new NoContentException("No hay granja con ID: " + id));
+	    }
+	
 	public void agregarGranja(Granja granja) {
-		granjaRepo.save(granja);
+		try {
+			granjaRepo.save(granja);
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
 	}
+	
+	public Granja editarGranja(Granja granja, Long id) {
+		   if (!granjaRepo.existsById(id)) {
+	            throw new NoContentException("No hay granja con ID: " + id);
+	        }
+	        granja.setId(id);
+	        return granjaRepo.save(granja);
+	    }
 
 	public void eliminarGranja(Long id) {
-		granjaRepo.deleteById(id);	
-	}
+		 if (!granjaRepo.existsById(id)) {
+	            throw new NoContentException("No hay granja con ID: " + id);
+	        }
+	        granjaRepo.deleteById(id);
+	    }	
+	
 	
 }
